requirements.txt
Flask==1.0.2
numpy==1.14.5
pandas==0.23.4
xlrd==1.1.0
XlsxWriter==1.2.1
openpyxl==3.0.0
mysql-connector-python==8.0.14
passlib==1.7.1
requests==2.20.0
biopython==1.73
nltk==3.4